Segregate AutoMockingTest
