+++ test/Tethos.FakeItEasy.Tests/AutoMockingTest/SUT/InheritedAutoMockingTest.cs
@@ -1,8 +1,8 @@
-﻿namespace Tethos.FakeItEasy.Tests.SUT
+﻿namespace Tethos.FakeItEasy.Tests.AutoMockingTest.SUT
 {
     using global::FakeItEasy;
 
-    public class InheritedAutoMockingTest : AutoMockingTest
+    public class InheritedAutoMockingTest : FakeItEasy.AutoMockingTest
     {
         public InheritedAutoMockingTest() => this.Proxy = this.Container = A.Fake<AutoFakeItEasyContainer>();
