+++ test/Tethos.NSubstitute.Tests/AutoMockingTest/AutoMockingTestTests.cs
@@ -1,0 +1,63 @@
+﻿namespace Tethos.NSubstitute.Tests.AutoMockingTest
+{
+    using System;
+    using AutoFixture.Xunit2;
+    using Castle.MicroKernel;
+    using Castle.MicroKernel.Registration;
+    using FluentAssertions;
+    using global::NSubstitute;
+    using Tethos.Extensions;
+    using Tethos.Tests.Common;
+    using Xunit;
+
+    public class AutoMockingTestTests : NSubstitute.AutoMockingTest
+    {
+        [Fact]
+        [Trait("Category", "Integration")]
+        public void Container_ShouldHaveAutoResolverInstalled()
+        {
+            // Assert
+            this.AutoResolver.Should().BeOfType<AutoNSubstituteResolver>();
+        }
+
+        [Theory]
+        [AutoData]
+        [Trait("Category", "Integration")]
+        public void Test_SimpleDependency_ShouldMatchValue(int expected)
+        {
+            // Arrange
+            var sut = this.Container.Resolve<SystemUnderTest>();
+            this.Container.Resolve<IMockable>()
+                .Do()
+                .Returns(expected);
+
+            // Act
+            var actual = sut.Do();
+
+            // Assert
+            actual.Should().Be(expected);
+        }
+
+        [Theory]
+        [AutoData]
+        [Trait("Category", "Integration")]
+        public void Clean_ShouldRevertBackToOriginalBehavior(Mockable mockable)
+        {
+            // Arrange
+            var sut = this.Container.Resolve<SystemUnderTest>();
+
+            this.Container.Register(Component.For<SystemUnderTest>()
+                .OverridesExistingRegistration()
+                .DependsOn(Dependency.OnValue<IMockable>(mockable)));
+
+            // Act
+            this.Clean();
+            var concrete = this.Container.Resolve<SystemUnderTest>();
+            Action action = () => concrete.Do();
+            sut.Do();
+
+            // Assert
+            action.Should().Throw<NotImplementedException>();
+        }
+    }
+}
