+++ test/Tethos.NSubstitute.Tests/AutoMockingTest/DependencyTests.cs
@@ -1,43 +1,15 @@
-﻿namespace Tethos.NSubstitute.Tests
+﻿namespace Tethos.NSubstitute.Tests.AutoMockingTest
 {
-    using System;
     using AutoFixture.Xunit2;
     using Castle.MicroKernel;
-    using Castle.MicroKernel.Registration;
     using FluentAssertions;
     using global::NSubstitute;
     using Tethos.Extensions;
     using Tethos.Tests.Common;
     using Xunit;
 
-    public class AutoMockingTestTests : AutoMockingTest
+    public class DependencyTests : NSubstitute.AutoMockingTest
     {
-        [Fact]
-        [Trait("Category", "Integration")]
-        public void Container_ShouldHaveAutoResolverInstalled()
-        {
-            // Assert
-            this.AutoResolver.Should().BeOfType<AutoNSubstituteResolver>();
-        }
-
-        [Theory]
-        [AutoData]
-        [Trait("Category", "Integration")]
-        public void Test_SimpleDependency_ShouldMatchValue(int expected)
-        {
-            // Arrange
-            var sut = this.Container.Resolve<SystemUnderTest>();
-            this.Container.Resolve<IMockable>()
-                .Do()
-                .Returns(expected);
-
-            // Act
-            var actual = sut.Do();
-
-            // Assert
-            actual.Should().Be(expected);
-        }
-
         [Fact]
         [Trait("Category", "Integration")]
         public void Container_Resolve_WithClassAndArguments_ShouldMockClass()
@@ -142,28 +114,6 @@
             this.Container.Resolve<Threshold>().Should().BeOfType(Substitute.For<Threshold>(true).GetType());
             this.Container.Resolve<PartialThreshold>().Should().BeOfType(Substitute.For<PartialThreshold>(true).GetType());
             this.Container.Resolve<AbstractThreshold>().Should().BeOfType(Substitute.For<AbstractThreshold>(true).GetType());
-        }
-
-        [Theory]
-        [AutoData]
-        [Trait("Category", "Integration")]
-        public void Clean_ShouldRevertBackToOriginalBehavior(Mockable mockable)
-        {
-            // Arrange
-            var sut = this.Container.Resolve<SystemUnderTest>();
-
-            this.Container.Register(Component.For<SystemUnderTest>()
-                .OverridesExistingRegistration()
-                .DependsOn(Dependency.OnValue<IMockable>(mockable)));
-
-            // Act
-            this.Clean();
-            var concrete = this.Container.Resolve<SystemUnderTest>();
-            Action action = () => concrete.Do();
-            sut.Do();
-
-            // Assert
-            action.Should().Throw<NotImplementedException>();
         }
     }
 }
